% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_who_line.R
\name{geom_sf_who_line}
\alias{geom_sf_who_line}
\title{wrapper function over geom_sf with defaults for showing adm0 country lines
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
geom_sf_who_line(
  mapping = aes(),
  data = NULL,
  col = who_map_col("border"),
  size = 0.2,
  ...
)
}
\arguments{
\item{mapping}{- passed down to mapping in geom_sf()}

\item{data}{- passed down to data in geom_sf()}

\item{col}{- passed down to col in geom_sf()}

\item{size}{- passed down to size in geom_sf()}
}
\description{
wrapper function over geom_sf with defaults for showing adm0 country lines
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
wrapper function to add all components to who map
}
