% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sf_who_poly.R
\name{geom_sf_who_poly}
\alias{geom_sf_who_poly}
\title{wrapper function over geom_sf with new defaults for no borders shown
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
geom_sf_who_poly(mapping = aes(), data = NULL, ..., col = NA)
}
\arguments{
\item{mapping}{- passed down to mapping in geom_sf()}

\item{data}{- passed down to data in geom_sf()}

\item{col}{- passed down to col in geom_sf()}
}
\description{
wrapper function over geom_sf with new defaults for no borders shown
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
wrapper function to add all components to who map
}
