% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_coords.R
\name{logo_coords}
\alias{logo_coords}
\title{function to add annotations to WHO map}
\usage{
logo_coords(
  location = c("topright", "bottomright", "topleft", "bottomleft"),
  outside_panel = TRUE
)
}
\arguments{
\item{location}{- user defined location - one of c("topright", "bottomright", "topleft", "bottomleft")}

\item{outside_panel}{- should the logo sit outside the panel? TRUE or FALSE}
}
\description{
function to add annotations to WHO map
}
\details{
function to determine coordinates for WHO logo based on predefined location
}
