% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_who_adm0.R
\name{pull_who_adm0}
\alias{pull_who_adm0}
\title{function to read SHP files for WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
pull_who_adm0()
}
\value{
a list of all shapefiles needed for mapping WHO maps at ADM0
}
\description{
function to read SHP files for WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to read SHP files for WHO map
}
