% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_who_adm1.R
\name{pull_who_adm1}
\alias{pull_who_adm1}
\title{function to read SHP files for WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
pull_who_adm1()
}
\value{
a list of all shapefiles needed for mapping WHO maps at ADM1
}
\description{
function to read SHP files for WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to read SHP files for WHO map ADM1 level
}
