% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_who_logo.R
\name{pull_who_logo}
\alias{pull_who_logo}
\title{function to read the WHO logo for mapping puroses
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
pull_who_logo(region = "HQ")
}
\arguments{
\item{region}{- which regional logo should be taken - defaults to HQ}
}
\value{
a PNG of the WHO logo
}
\description{
function to read the WHO logo for mapping puroses
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to read the WHO logo for mapping puroses
}
