% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_annotate.R
\name{who_map_annotate}
\alias{who_map_annotate}
\title{function to add annotations to WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_annotate(region = "HQ", data_source = "World Health Organization")
}
\arguments{
\item{region}{- user defined region for logo and disclaimer text}

\item{data_source}{- user defined data source}
}
\description{
function to add annotations to WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to add annotations to WHO map
}
