% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_col.R
\name{who_map_col}
\alias{who_map_col}
\title{function to get default who map colours
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_col(
  x = c("no_data", "not_applicable", "background", "lakes", "border", "title",
    "adm0_mask")
)
}
\arguments{
\item{x}{which colour index should be used}

\item{new_scale}{- should a new scale be added?}
}
\description{
function to get default who map colours
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to get default who map colours
}
