% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_legend_labs.R
\name{who_map_legend_labs}
\alias{who_map_legend_labs}
\title{code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_legend_labs(x)
}
\arguments{
\item{x}{- a label name}
}
\description{
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
helper function to convert label names
}
