% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_pipeline.R
\name{who_map_pipeline}
\alias{who_map_pipeline}
\title{wrapper function to add all components to who map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_pipeline(
  sf = pull_who_adm0(),
  xlim = NULL,
  ylim = NULL,
  region = "HQ",
  data_source = "World Health Organization",
  na_scale = TRUE,
  no_data_scale = TRUE
)
}
\arguments{
\item{sf}{- list of shapefiles to be used}

\item{xlim}{- the coordinate limits of the x axis}

\item{ylim}{- the coordinate limits of the y axis}

\item{region}{- user defined region for logo and disclaimer text}

\item{data_source}{- user defined data source}

\item{na_scale}{- should a new scale for not applicable be added?}

\item{add_no_data_scale}{- should a new scale for no data be added?}
}
\description{
wrapper function to add all components to who map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
wrapper function to add all components to who map
}
