% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_disp.R
\name{who_map_disp}
\alias{who_map_disp}
\title{function to add disputed territories to WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_disp(
  sf = pull_who_adm0(),
  na_scale = TRUE,
  no_data_scale = TRUE,
  background_col = who_map_col("background")
)
}
\arguments{
\item{sf}{list of shapefiles to be used}

\item{na_scale}{- should a new scale for not applicable be added?}

\item{no_data_scale}{- should a new scale for no data be added?}

\item{background_col}{- the colour of the background (ocean) and the lakes. Defaults to #E6E7E8}
}
\description{
function to add disputed territories to WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to add disputed territories to WHO map
}
