% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_disp.R
\name{who_map_disp}
\alias{who_map_disp}
\title{function to add disputed territories to WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_disp(sf = pull_who_adm0(), na_scale = TRUE, no_data_scale = TRUE)
}
\arguments{
\item{sf}{list of shapefiles to be used}

\item{na_scale}{- should a new scale for not applicable be added?}

\item{no_data_scale}{- should a new scale for no data be added?}
}
\description{
function to add disputed territories to WHO map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
function to add disputed territories to WHO map
}
