% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_pipeline.R
\name{who_map_pipeline}
\alias{who_map_pipeline}
\title{wrapper function to add all components to who map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_pipeline(sf = pull_who_adm0())
}
\arguments{
\item{sf}{- list of shapefiles to be used}
}
\description{
wrapper function to add all components to who map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
wrapper function to add all components to who map
}
