% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_pipeline.R
\name{who_map_pipeline}
\alias{who_map_pipeline}
\title{wrapper function to add all components to who map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_pipeline(
  sf = whomapper::pull_sfs(adm_level = 0, query_server = FALSE),
  xlim = NULL,
  ylim = NULL,
  box_lims = NULL,
  box_padding = 0.05,
  region = "HQ",
  data_source = "World Health Organization",
  production_team = "WHO Health Emergencies Programme",
  include_adm0_line = TRUE,
  na_scale = TRUE,
  no_data_scale = TRUE,
  no_annotation = FALSE,
  logo_location = c("topright", "bottomright", "topleft", "bottomleft"),
  logo_outside_panel = TRUE
)
}
\arguments{
\item{sf}{- list of shapefiles to be used}

\item{xlim}{- the coordinate limits of the x axis}

\item{ylim}{- the coordinate limits of the y axis}

\item{box_lims}{- coodrinate limits supplied as a bounding box generated by sf::st_bbox()}

\item{region}{- user defined region for logo and disclaimer text}

\item{data_source}{- user defined data source}

\item{production_team}{- for the caption, define the team that produced the map.
Defaults to _WHO Health Emergencies Programme_}

\item{include_adm0_line}{- should the specialised adm0 borders be used?
Note that these are borders **excluding** coastlines.}

\item{na_scale}{- should a new scale for not applicable be added?}

\item{no_annotation}{- if FALSE, no logo, title and annotation are appended.}

\item{logo_location}{- which part of the plot should the logo go?}

\item{add_no_data_scale}{- should a new scale for no data be added?}

\item{logo_outside_panel-}{should the logo be shown in the panel?}
}
\description{
wrapper function to add all components to who map
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
wrapper function to add all components to who map
}
