% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_map_theme.R
\name{who_map_theme}
\alias{who_map_theme}
\title{theme for WHO maps
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry}
\usage{
who_map_theme(xlim = NULL, ylim = NULL)
}
\arguments{
\item{xlim}{- the coordinate limits of the x axis}

\item{ylim}{- the coordinate limits of the y axis}
}
\description{
theme for WHO maps
code by IZAWA, Yurie & LAURENSON-SCHAFER Henry
}
\details{
theme for WHO maps
}
